fetcher/requirements.txt
pygrocy~=2.1.0
python-dotenv~=1.0.1
pyyaml~=6.0.2
requests~=2.32.3